requirements_ambilight.txt
siwat-light-control-protocol
pyautogui
scipy
numpy
pillow